refactor: add logo to search result for civic center service search
