+++ backend/src/modules/civic-center-service/interfaces/CivicCenterServiceFlat.ts
@@ -1,0 +1,7 @@
+import { CivicCenterService } from '../entities/civic-center-service.entity';
+
+export interface CivicCenterServiceFlat extends CivicCenterService {
+  organizationId: number;
+  organizationName: string;
+  logo: string;
+}
